requirements.txt
python-twitter==1.0
pymongo>=2.5.0